added a new state in the app js which saves the current centered map position, on every onclick which is going away from the map, the centered position will be saved
